+++ src/main/java/jsonvalues/spec/DecimalSchema.java
@@ -1,7 +1,6 @@
 package jsonvalues.spec;
 
 import java.math.BigDecimal;
-import java.util.Objects;
 
 /**
  * A class representing the schema for BigDecimal values in a JSON structure. It allows setting constraints such as
@@ -11,54 +10,30 @@ public final class DecimalSchema {
 
   private BigDecimal minimum;
   private BigDecimal maximum;
-  private BigDecimal multipleOf;
 
-  /**
-   * Sets the minimum value for BigDecimal numbers in the schema.
-   *
-   * @param minimum The minimum value (inclusive).
-   * @return This DecimalSchema instance for method chaining.
-   */
-  public DecimalSchema setMinimum(final BigDecimal minimum) {
-    this.minimum = Objects.requireNonNull(minimum);
-    return this;
+  public static DecimalSchema withMinimum(final BigDecimal minimum) {
+    var schema = new DecimalSchema();
+    schema.minimum = minimum;
+    return schema;
   }
 
-  /**
-   * Sets the maximum value for BigDecimal numbers in the schema.
-   *
-   * @param maximum The maximum value (inclusive).
-   * @return This DecimalSchema instance for method chaining.
-   */
-  public DecimalSchema setMaximum(final BigDecimal maximum) {
-    this.maximum = Objects.requireNonNull(maximum);
-    return this;
+  public static DecimalSchema withMaximum(final BigDecimal maximum) {
+    var schema = new DecimalSchema();
+    schema.maximum = maximum;
+    return schema;
   }
 
-  /**
-   * Sets a constraint for a multiple of BigDecimal numbers in the schema.
-   *
-   * @param multipleOf The value that the BigDecimal numbers must be a multiple of.
-   * @return This DecimalSchema instance for method chaining.
-   * @throws IllegalArgumentException If multipleOf is not a positive number.
-   */
-  public DecimalSchema setMultipleOf(final BigDecimal multipleOf) {
-    if (Objects.requireNonNull(multipleOf)
-               .signum() <= 0) {
-      throw new IllegalArgumentException("multipleOf must be a positive number");
-    }
-    this.multipleOf = multipleOf;
-    return this;
+  public static DecimalSchema betweenInterval(final BigDecimal minimum,
+                                              final BigDecimal maximum) {
+    var schema = new DecimalSchema();
+    schema.minimum = minimum;
+    schema.maximum = maximum;
+    return schema;
   }
 
-  /**
-   * Builds and returns an instance of DecimalSchemaConstraints based on the specified constraints.
-   *
-   * @return An instance of DecimalSchemaConstraints with the specified constraints.
-   */
+
   DecimalSchemaConstraints build() {
     return new DecimalSchemaConstraints(minimum,
-                                        maximum,
-                                        multipleOf);
+                                        maximum);
   }
 }